+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - The `CellularAutomata1D` model plugin has been updated to implement the 256 elementary cellular automaton rules
+- SingleValue - `min()`, `max()` and `rand()` now return an invalid Value
 
 ### Fixed
 - Fixes #27 - Experiment Designer: vertical scrollbar is hiding the buttons and fields